Move make_lock_spec from conda_lock.py to src_parser/__init__.py
